requirements.txt
# Flask Core
Flask==3.0.0
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3

# Database
pymongo==4.6.1
dnspython==2.4.2

# Security
bcrypt==4.1.2
WTForms==3.1.1
email-validator==2.1.0

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Additional
Werkzeug==3.0.1
